run as alias for app-run
